+++ Package.swift
@@ -10,7 +10,7 @@ var sources = [
     "ggml/src/ggml-alloc.c",
     "ggml/src/ggml-backend.c",
     "ggml/src/ggml-quants.c",
-    "ggml/src/ggml-aarch64.cpp",
+    "ggml/src/ggml-aarch64.c",
 ]
 
 var resources: [Resource] = []
